Portable scripting tests
