+++ .nuget/V3/packages.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="System.Dynamic.Runtime" version="4.0.10" />
+  <package id="Microsoft.CSharp" version="4.0.0-beta-22816" />
+  <package id="System.Reflection.Emit.Lightweight" version="4.0.0-beta-22816" />
+  <package id="System.Linq.Expressions" version="4.0.10-beta-22816" />
+</packages>
