+++ .nuget/packages.config
@@ -1,5 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright (c) Microsoft Open Technologies, Inc.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information. -->
 <packages>
   <package id="Microsoft.Net.Compilers" version="1.1.0-beta1-20150727-01" targetFramework="net45" />
   <package id="Microsoft.Net.RoslynDiagnostics" version="1.1.0-beta1-20150727-01" targetFramework="net45" />
@@ -11,7 +10,36 @@
   <package id="dnx-coreclr-win-x86" version="1.0.0-beta5-12101" />
   <package id="dnx-mono" version="1.0.0-beta5-12101" />
   <package id="Microsoft.Build.Mono.Debug" version="14.1.0.0-prerelease" />
+  <!-- FX -->
+  <package id="System.IO" version="4.0.10-beta-22816" />
+  <package id="System.IO.FileSystem" version="4.0.0-beta-22816" />
+  <package id="System.IO.FileSystem.Primitives" version="4.0.0-beta-22816" />
+  <package id="System.Runtime" version="4.0.20-beta-22816" />
+  <package id="System.Runtime.Extensions" version="4.0.10-beta-22816" />
+  <package id="System.Runtime.InteropServices" version="4.0.20-beta-22816" />
+  <package id="System.Runtime.Loader" version="4.0.0-beta-22816" />
+  <package id="System.Text.RegularExpressions" version="4.0.10-beta-22816" />
+  <package id="System.Reflection" version="4.0.10-beta-22816" />
+  <package id="System.Reflection.Extensions" version="4.0.0-beta-22816" />
+  <package id="System.Reflection.Primitives" version="4.0.0-beta-22816" />
+  <package id="System.Resources.ResourceManager" version="4.0.0-beta-22816" />
+  <package id="System.Collections" version="4.0.10-beta-22816" />
+  <package id="System.Collections.Concurrent" version="4.0.10-beta-22816" />
+  <package id="System.Linq" version="4.0.0-beta-22816" />
+  <package id="System.Diagnostics.Debug" version="4.0.10-beta-22816" />
+  <package id="System.Diagnostics.Tools" version="4.0.0-beta-22816" />
+  <package id="System.Globalization" version="4.0.10-beta-22816" />
+  <package id="System.Text.Encoding" version="4.0.10-beta-22816" />
+  <package id="System.Threading" version="4.0.10-beta-22816" />
+  <package id="System.Threading.Tasks" version="4.0.10-beta-22816" />
+  <package id="System.Threading.Thread" version="4.0.0-beta-22816" />
+  <package id="System.Xml.XDocument" version="4.0.10-beta-22816" />
+  <package id="System.Xml.XmlDocument" version="4.0.0-beta-22816" />
+  <package id="System.IO.MemoryMappedFiles" version="4.0.0-beta-22816" />
+  <package id="System.Console" version="4.0.0-beta-22816" />
+  <package id="System.Xml.ReaderWriter" version="4.0.10-beta-22816" />
   <!-- xunit -->
+  <package id="xunit" version="1.9.2" targetFramework="net45" />
   <package id="xunit" version="2.1.0-beta4-build3109" />
   <package id="xunit.abstractions" version="2.0.0" />
   <package id="xunit.assert" version="2.1.0-beta4-build3109" />
