+++ src/EditorFeatures/CSharpTest/CSharpEditorServicesTest.csproj
@@ -97,7 +97,7 @@
       <Project>{76C6F005-C89D-4348-BB4A-39189DDBEB52}</Project>
       <Name>ServicesTestUtilities</Name>
     </ProjectReference>
-    <ProjectReference Include="..\..\Scripting\Test\ScriptingTest.csproj">
+    <ProjectReference Include="..\..\Scripting\CoreTest\ScriptingTest.csproj">
       <Project>{2DAE4406-7A89-4B5F-95C3-BC5472CE47CE}</Project>
       <Name>ScriptingTest</Name>
       <ReferenceOutputAssembly>false</ReferenceOutputAssembly>
