+++ src/ExpressionEvaluator/CSharp/Test/ResultProvider/CSharpResultProviderTest.csproj
@@ -50,7 +50,7 @@
       <Project>{76C6F005-C89D-4348-BB4A-391898DBEB52}</Project>
       <Name>TestUtilities.Desktop</Name>
     </ProjectReference>
-    <ProjectReference Include="..\..\..\..\Scripting\Test\ScriptingTest.csproj">
+    <ProjectReference Include="..\..\..\..\Scripting\CoreTest\ScriptingTest.csproj">
       <Project>{2DAE4406-7A89-4B5F-95C3-BC5472CE47CE}</Project>
       <Name>ScriptingTest</Name>
       <ReferenceOutputAssembly>false</ReferenceOutputAssembly>
