+++ src/Scripting/CSharpTest/InteractiveSessionFixtures.cs
@@ -1,0 +1,24 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+public class InteractiveFixtures_TopLevelHostObject
+{
+    public int X, Y, Z;
+}
+
+namespace InteractiveFixtures
+{
+    namespace A
+    {
+        public class X { }
+    }
+
+    namespace B
+    {
+        public class X { }
+    }
+
+    namespace C
+    {
+        public interface System { }
+    }
+}
