+++ src/Scripting/CSharpTest/ObjectFormatterTests.cs
@@ -14,7 +14,6 @@ using Microsoft.CodeAnalysis.Scripting.UnitTests;
 using ObjectFormatterFixtures;
 using Roslyn.Test.Utilities;
 using Xunit;
-using VB = Microsoft.CodeAnalysis.VisualBasic;
 
 namespace Microsoft.CodeAnalysis.Scripting.CSharp.UnitTests
 {
@@ -469,15 +468,6 @@ namespace Microsoft.CodeAnalysis.Scripting.CSharp.UnitTests
         }
 
         [Fact]
-        public void DebuggerProxy_FrameworkTypes_ArrayList()
-        {
-            var obj = new ArrayList { 1, 2, true, "foo" };
-            var str = CSharpObjectFormatter.Instance.FormatObject(obj, s_inline);
-
-            Assert.Equal("ArrayList(4) { 1, 2, true, \"foo\" }", str);
-        }
-
-        [Fact]
         public void DebuggerProxy_FrameworkTypes_BitArray()
         {
             // BitArray doesn't have debugger proxy/display
@@ -487,42 +477,27 @@ namespace Microsoft.CodeAnalysis.Scripting.CSharp.UnitTests
         }
 
         [Fact]
-        public void DebuggerProxy_FrameworkTypes_Hashtable()
-        {
-            var obj = new Hashtable
-            {
-                { new byte[] { 1, 2 }, new[] { 1,2,3 } },
-            };
-
-            var str = CSharpObjectFormatter.Instance.FormatObject(obj, s_memberList);
-
-            AssertMembers(str, "Hashtable(1)",
-                "{ byte[2] { 1, 2 }, int[3] { 1, 2, 3 } }"
-            );
-        }
-
-        [Fact]
         public void DebuggerProxy_FrameworkTypes_Queue()
         {
-            var obj = new Queue();
+            var obj = new Queue<int>();
             obj.Enqueue(1);
             obj.Enqueue(2);
             obj.Enqueue(3);
 
             var str = CSharpObjectFormatter.Instance.FormatObject(obj, s_inline);
-            Assert.Equal("Queue(3) { 1, 2, 3 }", str);
+            Assert.Equal("Queue<int>(3) { 1, 2, 3 }", str);
         }
 
         [Fact]
         public void DebuggerProxy_FrameworkTypes_Stack()
         {
-            var obj = new Stack();
+            var obj = new Stack<int>();
             obj.Push(1);
             obj.Push(2);
             obj.Push(3);
 
             var str = CSharpObjectFormatter.Instance.FormatObject(obj, s_inline);
-            Assert.Equal("Stack(3) { 3, 2, 1 }", str);
+            Assert.Equal("Stack<int>(3) { 3, 2, 1 }", str);
         }
 
         [Fact]
@@ -577,19 +552,19 @@ namespace Microsoft.CodeAnalysis.Scripting.CSharp.UnitTests
         [Fact]
         public void DebuggerProxy_FrameworkTypes_SortedList()
         {
-            SortedList obj = new SortedList();
+            var obj = new SortedList<int, int>();
             obj.Add(3, 4);
             obj.Add(1, 5);
             obj.Add(2, 6);
 
             var str = CSharpObjectFormatter.Instance.FormatObject(obj, s_inline);
-            Assert.Equal("SortedList(3) { { 1, 5 }, { 2, 6 }, { 3, 4 } }", str);
+            Assert.Equal("SortedList<int, int>(3) { { 1, 5 }, { 2, 6 }, { 3, 4 } }", str);
 
-            obj = new SortedList();
-            obj.Add(new[] { 3 }, new int[] { 4 });
+            var obj2 = new SortedList<int[], int[]>();
+            obj2.Add(new[] { 3 }, new int[] { 4 });
 
-            str = CSharpObjectFormatter.Instance.FormatObject(obj, s_inline);
-            Assert.Equal("SortedList(1) { { int[1] { 3 }, int[1] { 4 } } }", str);
+            str = CSharpObjectFormatter.Instance.FormatObject(obj2, s_inline);
+            Assert.Equal("SortedList<int[], int[]>(1) { { int[1] { 3 }, int[1] { 4 } } }", str);
         }
 
         [Fact]
@@ -806,49 +781,6 @@ namespace Microsoft.CodeAnalysis.Scripting.CSharp.UnitTests
                  "4",
                  "5"
             );
-        }
-
-        [Fact]
-        public void VBBackingFields_DebuggerBrowsable()
-        {
-            string source = @"
-Imports System
-
-Class C
-   Public WithEvents WE As C
-   Public Event E As Action
-   Public Property A As Integer
-End Class
-";
-            var compilation = VB.VisualBasicCompilation.Create(
-                "foo",
-                new[] { VB.VisualBasicSyntaxTree.ParseText(source) },
-                new[] { MetadataReference.CreateFromAssemblyInternal(typeof(object).Assembly) },
-                new VB.VisualBasicCompilationOptions(OutputKind.DynamicallyLinkedLibrary, optimizationLevel: OptimizationLevel.Debug));
-
-            Assembly a;
-            using (var stream = new MemoryStream())
-            {
-                var result = compilation.Emit(stream);
-                a = Assembly.Load(stream.ToArray());
-            }
-
-            var c = a.GetType("C");
-            var obj = Activator.CreateInstance(c);
-
-            var str = CSharpObjectFormatter.Instance.FormatObject(obj, s_memberList);
-            AssertMembers(str, "C",
-                "A: 0",
-                "WE: null"
-            );
-
-            var attrsA = c.GetField("_A", BindingFlags.Instance | BindingFlags.NonPublic).GetCustomAttributes(typeof(DebuggerBrowsableAttribute), true);
-            var attrsWE = c.GetField("_WE", BindingFlags.Instance | BindingFlags.NonPublic).GetCustomAttributes(typeof(DebuggerBrowsableAttribute), true);
-            var attrsE = c.GetField("EEvent", BindingFlags.Instance | BindingFlags.NonPublic).GetCustomAttributes(typeof(DebuggerBrowsableAttribute), true);
-
-            Assert.Equal(1, attrsA.Length);
-            Assert.Equal(1, attrsWE.Length);
-            Assert.Equal(1, attrsE.Length);
         }
     }
 }
