+++ src/Scripting/CoreTest.Desktop/GlobalAssemblyCacheTests.cs
