+++ src/Scripting/CoreTest.Desktop/MetadataShadowCopyProviderTests.cs
@@ -1,5 +1,4 @@
 // Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
 using System;
 using System.Linq;
 using System.Collections.Generic;
@@ -15,6 +14,8 @@ using System.Runtime.InteropServices;
 
 namespace Roslyn.Services.UnitTests
 {
+    // TODO: clean up and move to portable tests
+
     public class MetadataShadowCopyProviderTests : TestBase, IDisposable
     {
         private readonly MetadataShadowCopyProvider _provider;
