+++ src/Scripting/CoreTest/ObjectFormatterTestBase.cs
