+++ src/Scripting/CoreTest/ObjectFormatterTests.Fixtures.cs
