+++ src/Scripting/CoreTest/ScriptingTestHelpers.cs
