+++ src/Scripting/CoreTest/Session.cs
