--- src/Scripting/VisualBasic/packages.config
@@ -1,22 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="System.IO" version="4.0.10-beta-22816" />
-  <package id="System.IO.FileSystem" version="4.0.0-beta-22816" />
-  <package id="System.IO.FileSystem.Primitives" version="4.0.0-beta-22816" />
-  <package id="System.Runtime" version="4.0.20-beta-22816" />
-  <package id="System.Runtime.Extensions" version="4.0.10-beta-22816" />
-  <package id="System.Runtime.InteropServices" version="4.0.20-beta-22816" />
-  <package id="System.Runtime.Loader" version="4.0.0-beta-22816" />
-  <package id="System.Reflection" version="4.0.10-beta-22816" />
-  <package id="System.Reflection.Extensions" version="4.0.0-beta-22816" />
-  <package id="System.Reflection.Primitives" version="4.0.0-beta-22816" />
-  <package id="System.Resources.ResourceManager" version="4.0.0-beta-22816" />
-  <package id="System.Collections" version="4.0.10-beta-22816" />
-  <package id="System.Linq" version="4.0.0-beta-22816" />
-  <package id="System.Diagnostics.Debug" version="4.0.10-beta-22816" />
-  <package id="System.Diagnostics.Tools" version="4.0.0-beta-22816" />
-  <package id="System.Globalization" version="4.0.10-beta-22816" />
-  <package id="System.Text.Encoding" version="4.0.10-beta-22816" />
-  <package id="System.Threading" version="4.0.10-beta-22816" />
-  <package id="System.Threading.Tasks" version="4.0.10-beta-22816" />
-</packages>
