+++ src/Scripting/VisualBasicTest.Desktop/InteractiveSessionTests.vb
@@ -1,0 +1,15 @@
+' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+Imports Microsoft.CodeAnalysis.Scripting.Test
+Imports Microsoft.CodeAnalysis.VisualBasic
+Imports Roslyn.Test.Utilities
+Imports Xunit
+
+Namespace Microsoft.CodeAnalysis.Scripting.VisualBasic.UnitTests
+
+    Public Class InteractiveSessionTests
+        Inherits TestBase
+
+        ' TODO: port tests from C#
+    End Class
+End Namespace
