+++ src/Scripting/VisualBasicTest/ObjectFormatterTests.vb
@@ -66,6 +66,7 @@ Namespace Microsoft.CodeAnalysis.Scripting.VisualBasic.UnitTests
             str = VisualBasicObjectFormatter.Instance.FormatObject(sort, New ObjectFormattingOptions(maxLineLength:=80, memberFormat:=MemberDisplayFormat.Inline))
             Assert.Equal("Sort { aB=-1, ab=1, Ac=-1, Ad=1, ad=-1, aE=1, aF=-1, AG=1 }", str)
         End Sub
-    End Class
 
+        ' TODO: port tests from C#
+    End Class
 End Namespace
