+++ src/Scripting/VisualBasicTest/VisualBasicScriptEngine.vb
@@ -1,6 +1,5 @@
 ' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-Imports Microsoft.CodeAnalysis.VisualBasic.UnitTests
 Imports Roslyn.Test.Utilities
 
 Namespace Microsoft.CodeAnalysis.Scripting.VisualBasic
