+++ src/Test/Utilities/Portable/FX/EnsureEnglishUICulture.cs
@@ -37,8 +37,11 @@ namespace Roslyn.Test.Utilities
                 _threadUICulture = CultureInfo.CurrentUICulture;
                 _needToRestore = true;
 
-                // TODO: Should use CultureInfo.CurrentUICulture on 4.6
+#if DNX
+                CultureInfo.CurrentUICulture = preferred;
+#else
                 System.Threading.Thread.CurrentThread.CurrentUICulture = preferred;
+#endif
             }
         }
 
@@ -49,8 +52,11 @@ namespace Roslyn.Test.Utilities
             if (_needToRestore && _threadId == Thread.CurrentThread.ManagedThreadId)
             {
                 _needToRestore = false;
-                // TODO: Should use CultureInfo.CurrentUICulture on 4.6
+#if DNX
+                CultureInfo.CurrentUICulture = _threadUICulture;
+#else
                 System.Threading.Thread.CurrentThread.CurrentUICulture = _threadUICulture;
+#endif
             }
         }
     }
